Changed the return type of DropChance#getChance from an int to a double

+++ src/main/java/me/swanis/mobcoins/chance/DropChance.java
@@ -2,6 +2,6 @@ package me.swanis.mobcoins.chance;
 
 public abstract class DropChance {
 
-    public abstract int getChance();
+    public abstract double getChance();
 
 }
